+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `PngRenderer` now correctly respects `invert` option (#32)
+- `SvgRenderer` now correctly inverts module pattern when `invert` option is enabled (#32)
 
 ## [0.4.10] - 2026-03-17